# api/openapi.yaml
openapi: 3.1.0
info:
  title: DevOps Incident Platform API
  version: 0.0.1
  description: REST API for the DevOps incident management platform.
  license:
    name: MIT
    url: https://opensource.org/license/mit
...
servers:
  - url: http://localhost:4010
    description: Local mock server (pixi run mock-api)
  - url: http://localhost:8080/api/v1
    description: Local gateway and Swagger UI (pixi run compose-up, same paths as Helm)
  - url: /api/v1
    description: API gateway (relative, in-cluster)
paths:
...
  /auth/login:
    post:
      operationId: loginUser
      summary: Authenticate and start a session
      description: |
        On success, user-service sets an httpOnly `session` cookie containing a signed JWT
        (`SameSite=Strict`). The gateway validates this cookie and injects `X-User-Id` /
        `X-User-Role` on downstream requests. Do not send credentials on subsequent API calls.
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated
          headers:
            Set-Cookie:
              description: httpOnly session JWT (`session` cookie)
              schema:
                type: string
                example: session=eyJhbG...; Path=/; HttpOnly; SameSite=Strict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /users/me:
    get:
      operationId: getCurrentUser
      summary: Get the authenticated user profile
      description: |
        Requires a valid `session` cookie (browser) or prior login via `POST /auth/login` on the same HTTP client.
        Python callers should use `Client` and pass cookies (e.g. `client.with_cookies({"session": "<jwt>"})`), not
        `AuthenticatedClient` / `Authorization: Bearer`.
      tags: [users]
      security: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCurrentUser
      summary: Update the authenticated user profile
      description: |
        Partial update of the caller's own profile. Requires a valid `session` cookie (see
        `GET /users/me`). Changing `email` additionally requires `currentPassword`; changing
        only `displayName` does not. The session cookie stays valid after the update.
      tags: [users]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid request (validation failed, or email change without currentPassword)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated, or currentPassword is wrong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already registered to another account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/me/password:
    post:
      operationId: changePassword
      summary: Change the authenticated user's password
      description: |
        Requires a valid `session` cookie (see `GET /users/me`) and the current password.
        Existing sessions are not revoked (ADR 0007: stateless session JWTs expire via TTL).
      tags: [users]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "204":
          description: Password changed
        "400":
          description: Invalid request (validation failed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated, or currentPassword is wrong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users:
    get:
      operationId: listUsers
      summary: List users (directory)
      description: |
        Returns all users for assignment pickers and mention lists. Requires a valid `session` cookie (see
        `GET /users/me`). Python callers use `Client` + cookies, not Bearer auth.
      tags: [users]
      security: []
      parameters:
        - $ref: "#/components/parameters/UserListLimitParam"
        - $ref: "#/components/parameters/UserListOffsetParam"
      responses:
        "200":
          description: User directory page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
    UserListLimitParam:
      name: limit
      in: query
      required: false
      description: Maximum number of users to return (default 50, max 100).
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    UserListOffsetParam:
      name: offset
      in: query
      required: false
      description: Number of users to skip for pagination.
      schema:
        type: integer
        minimum: 0
        default: 0
...
  schemas:
    UserRole:
      type: string
      description: |
        Platform role aligned with the analysis object model.
        `MEMBER` maps to TeamMember, `RESPONDER` to IncidentResponder, `COMMANDER` to IncidentCommander.
      enum: [MEMBER, RESPONDER, COMMANDER]
      example: MEMBER
    User:
      type: object
      description: Public user profile (no credentials).
      required: [id, email, displayName, role, createdAt]
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
          example: 018e2c5f-1234-7abc-8def-0000000000aa
        email:
          type: string
          format: email
          example: responder@example.com
        displayName:
          type: string
          minLength: 1
          maxLength: 100
          example: Alex Responder
        role:
          $ref: "#/components/schemas/UserRole"
        createdAt:
          type: string
          format: date-time
          example: "2026-05-08T10:00:00Z"
...
    UpdateProfileRequest:
      type: object
      description: >-
        Partial profile update; at least one of `email` or `displayName` must be
        present. `currentPassword` is required when `email` is present.
      additionalProperties: false
      minProperties: 1
      properties:
        email:
          type: string
          format: email
          example: new.address@example.com
        displayName:
          type: string
          minLength: 1
          maxLength: 100
          example: Alex Responder
        currentPassword:
          type: string
          format: password
          writeOnly: true
          example: "change-me-8+"
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      additionalProperties: false
      properties:
        currentPassword:
          type: string
          format: password
          writeOnly: true
          example: "change-me-8+"
        newPassword:
          type: string
          format: password
          minLength: 8
          maxLength: 128
          writeOnly: true
          example: "new-secret-9!"
    LoginRequest:
      type: object
      required: [email, password]
      additionalProperties: false
      properties:
        email:
          type: string
          format: email
          example: responder@example.com
        password:
          type: string
          format: password
          writeOnly: true
          example: "change-me-8+"
          minLength: 8
    UserListResponse:
      type: object
      required: [items, total, limit, offset]
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
          minimum: 0
          description: Total users matching the query (ignoring pagination).
          example: 2
        limit:
          type: integer
          minimum: 1
          maximum: 100
          example: 50
        offset:
          type: integer
          minimum: 0
          example: 0
    ErrorResponse:
      type: object
      description: Simple error payload for auth and validation failures.
      required: [message]
      additionalProperties: false
      properties:
        message:
          type: string
          example: Invalid email or password
...