# services/generated/java/src/main/java/org/openapitools/api/UsersApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.14.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.ChangePasswordRequest;
import org.openapitools.model.ErrorResponse;
import org.openapitools.model.UpdateProfileRequest;
import org.openapitools.model.User;
import org.openapitools.model.UserListResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.lang.Nullable;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.14.0")
@Validated
@Tag(name = "users", description = "User profiles and directory (user-service). Runtime auth uses the `session` cookie or gateway-injected `X-User-Id` / `X-User-Role`; OpenAPI marks operations `security: []` for codegen only. ")
public interface UsersApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /users/me/password : Change the authenticated user&#39;s password
     * Requires a valid &#x60;session&#x60; cookie (see &#x60;GET /users/me&#x60;) and the current password. Existing sessions are not revoked (ADR 0007: stateless session JWTs expire via TTL).
     *
     * @param changePasswordRequest  (required)
     * @return Password changed (status code 204)
     *         or Invalid request (validation failed) (status code 400)
     *         or Not authenticated, or currentPassword is wrong (status code 401)
     */
    @Operation(
        operationId = "changePassword",
        summary = "Change the authenticated user's password",
        description = "Requires a valid `session` cookie (see `GET /users/me`) and the current password. Existing sessions are not revoked (ADR 0007: stateless session JWTs expire via TTL). ",
        tags = { "users" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Password changed"),
            @ApiResponse(responseCode = "400", description = "Invalid request (validation failed)", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Not authenticated, or currentPassword is wrong", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/users/me/password",
        produces = { "application/json" },
        consumes = { "application/json" }
    )

    default ResponseEntity<Void> changePassword(
        @Parameter(name = "ChangePasswordRequest", description = "", required = true) @Valid @RequestBody ChangePasswordRequest changePasswordRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"Invalid email or password\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"Invalid email or password\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /users/me : Get the authenticated user profile
     * Requires a valid &#x60;session&#x60; cookie (browser) or prior login via &#x60;POST /auth/login&#x60; on the same HTTP client. Python callers should use &#x60;Client&#x60; and pass cookies (e.g. &#x60;client.with_cookies({\&quot;session\&quot;: \&quot;&lt;jwt&gt;\&quot;})&#x60;), not &#x60;AuthenticatedClient&#x60; / &#x60;Authorization: Bearer&#x60;.
     *
     * @return Current user (status code 200)
     *         or Not authenticated (status code 401)
     */
    @Operation(
        operationId = "getCurrentUser",
        summary = "Get the authenticated user profile",
        description = "Requires a valid `session` cookie (browser) or prior login via `POST /auth/login` on the same HTTP client. Python callers should use `Client` and pass cookies (e.g. `client.with_cookies({\"session\": \"<jwt>\"})`), not `AuthenticatedClient` / `Authorization: Bearer`. ",
        tags = { "users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Current user", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = User.class))
            }),
            @ApiResponse(responseCode = "401", description = "Not authenticated", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/users/me",
        produces = { "application/json" }
    )

    default ResponseEntity<User> getCurrentUser(

    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"createdAt\" : \"2026-05-08T10:00:00Z\", \"role\" : \"MEMBER\", \"displayName\" : \"Alex Responder\", \"id\" : \"018e2c5f-1234-7abc-8def-0000000000aa\", \"email\" : \"responder@example.com\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"Invalid email or password\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /users : List users (directory)
     * Returns all users for assignment pickers and mention lists. Requires a valid &#x60;session&#x60; cookie (see &#x60;GET /users/me&#x60;). Python callers use &#x60;Client&#x60; + cookies, not Bearer auth.
     *
     * @param limit Maximum number of users to return (default 50, max 100). (optional, default to 50)
     * @param offset Number of users to skip for pagination. (optional, default to 0)
     * @return User directory page (status code 200)
     *         or Not authenticated (status code 401)
     */
    @Operation(
        operationId = "listUsers",
        summary = "List users (directory)",
        description = "Returns all users for assignment pickers and mention lists. Requires a valid `session` cookie (see `GET /users/me`). Python callers use `Client` + cookies, not Bearer auth. ",
        tags = { "users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "User directory page", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserListResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Not authenticated", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/users",
        produces = { "application/json" }
    )

    default ResponseEntity<UserListResponse> listUsers(
        @Min(1) @Max(100) @Parameter(name = "limit", description = "Maximum number of users to return (default 50, max 100).", in = ParameterIn.QUERY) @Valid @RequestParam(value = "limit", required = false, defaultValue = "50") Integer limit,
        @Min(0) @Parameter(name = "offset", description = "Number of users to skip for pagination.", in = ParameterIn.QUERY) @Valid @RequestParam(value = "offset", required = false, defaultValue = "0") Integer offset
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"total\" : 2, \"offset\" : 0, \"limit\" : 50, \"items\" : [ { \"createdAt\" : \"2026-05-08T10:00:00Z\", \"role\" : \"MEMBER\", \"displayName\" : \"Alex Responder\", \"id\" : \"018e2c5f-1234-7abc-8def-0000000000aa\", \"email\" : \"responder@example.com\" }, { \"createdAt\" : \"2026-05-08T10:00:00Z\", \"role\" : \"MEMBER\", \"displayName\" : \"Alex Responder\", \"id\" : \"018e2c5f-1234-7abc-8def-0000000000aa\", \"email\" : \"responder@example.com\" } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"Invalid email or password\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PATCH /users/me : Update the authenticated user profile
     * Partial update of the caller&#39;s own profile. Requires a valid &#x60;session&#x60; cookie (see &#x60;GET /users/me&#x60;). Changing &#x60;email&#x60; additionally requires &#x60;currentPassword&#x60;; changing only &#x60;displayName&#x60; does not. The session cookie stays valid after the update.
     *
     * @param updateProfileRequest  (required)
     * @return Updated user profile (status code 200)
     *         or Invalid request (validation failed, or email change without currentPassword) (status code 400)
     *         or Not authenticated, or currentPassword is wrong (status code 401)
     *         or Email already registered to another account (status code 409)
     */
    @Operation(
        operationId = "updateCurrentUser",
        summary = "Update the authenticated user profile",
        description = "Partial update of the caller's own profile. Requires a valid `session` cookie (see `GET /users/me`). Changing `email` additionally requires `currentPassword`; changing only `displayName` does not. The session cookie stays valid after the update. ",
        tags = { "users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Updated user profile", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = User.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request (validation failed, or email change without currentPassword)", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Not authenticated, or currentPassword is wrong", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "409", description = "Email already registered to another account", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PATCH,
        value = "/users/me",
        produces = { "application/json" },
        consumes = { "application/json" }
    )

    default ResponseEntity<User> updateCurrentUser(
        @Parameter(name = "UpdateProfileRequest", description = "", required = true) @Valid @RequestBody UpdateProfileRequest updateProfileRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"createdAt\" : \"2026-05-08T10:00:00Z\", \"role\" : \"MEMBER\", \"displayName\" : \"Alex Responder\", \"id\" : \"018e2c5f-1234-7abc-8def-0000000000aa\", \"email\" : \"responder@example.com\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"Invalid email or password\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"Invalid email or password\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"Invalid email or password\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}